server/deploy/pubsub-demo/scripts/client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = "Kafka clients for testing this demo deployment"
authors = ["Kacper Zienkiewicz <kzienkiewicz@internships.antmicro.com>"]
license = "Apache 2.0"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
kafka-python = "==2.1.5"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
